split object declarations into separate classes
